requirements.txt
streamlit
audio-recorder-streamlit
openai
deepgram-sdk
elevenlabs